server change

+++ server.c
@@ -48,7 +48,7 @@ int main(int argc, char *argv[])
 
 	struct sockaddr_in sock_var;
 	int serverFileDescriptor = socket(AF_INET,SOCK_STREAM,0);
-	int clientFileDescriptor;
+	long clientFileDescriptor;
 	int i;
 	pthread_t t[1000];
 
@@ -65,7 +65,7 @@ int main(int argc, char *argv[])
 			{
 				clientFileDescriptor=accept(serverFileDescriptor,NULL,NULL);
 				printf("nConnected to client %d\n",clientFileDescriptor);
-				pthread_create(&t[i],NULL,clientThreadHandler,(void *)clientFileDescriptor);
+				pthread_create(&t[i],NULL,clientThreadHandler,(void *) clientFileDescriptor);
 			}
 		}
 		close(serverFileDescriptor);
